change module dir
