+++ lib/kakine.rb
@@ -1,10 +1,10 @@
 require "kakine/version"
 require 'kakine/cli'
+require 'kakine/cli/operation'
 require 'kakine/adapter'
 require 'kakine/resource'
-require 'kakine/operation'
 require 'kakine/security_group'
-require 'kakine/diff_parser'
+require 'kakine/security_group/diff_parser'
 
 module Kakine
 end
