+++ lib/kakine/security_group.rb
@@ -1,13 +1,14 @@
-require 'kakine/diff_parser'
+require 'kakine/security_group/diff_parser'
 module Kakine
   class SecurityGroup
     attr_reader :description, :tenant_name
     include DiffParser
 
     def initialize(tenant_name, diff)
+      unset_security_rules
       @diff = diff
       @registered_sg = Kakine::Resource.security_groups_hash(tenant_name)
-      unset_security_rules
+
       init_parse_diff
       set_remote_security_group_id
